backend/requirements.txt
# ATVISION Backend Dependencies

# CUDA-enabled PyTorch (for GPU acceleration, Python 3.11 compatible)
--extra-index-url https://download.pytorch.org/whl/cu126
torch 
torchvision
# USE triton-windows FOR WINDOWS SYSTEMS
triton-windows
# USE triton FOR LINUX/MAC SYSTEMS
# triton

# YOLO and Computer Vision
# faiss-gpu-cu12
ultralytics>=8.4.6
opencv-python>=4.8.0
numpy>=1.24.0

# SAM3 (Segment Anything Model 3) - VisionMask
# Install from source: pip install git+https://github.com/facebookresearch/sam3.git
# Or local installation if available
sam3

# FastAPI and server
fastapi>=0.104.0
uvicorn[standard]>=0.24.0

# Database
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.9
alembic>=1.12.0
pandas

# Pydantic
pydantic>=2.5.0
pydantic-settings>=2.1.0
pydantic[email]>=2.3.0

# Authentication
python-jose[cryptography]>=3.3.0
bcrypt>=4.0.0

# File handling
python-multipart>=0.0.6
pillow>=10.1.0
aiofiles>=23.2.1

# Utilities
python-dotenv>=1.0.0
psutil>=5.9.0

# Export functionality
reportlab
matplotlab

# Workflow dependencies
simpleeval>=0.9.13
apscheduler>=3.10.4
requests>=2.31.0
jinja2>=3.1.2

# Recognition Catalog - CLIP embeddings and vector similarity
transformers>=4.35.0
sentence-transformers>=2.2.2
pgvector>=0.2.4
